Change chunk compression

+++ src/world.rs
@@ -61,7 +61,7 @@ impl World {
     }
 
     /// Return data at position x,y,z
-    pub fn _get_data(&self, x: i64, y: i64, z: i64) -> u32 {
+    pub fn _get_data(&self, x: i64, y: i64, z: i64) -> u16 {
         let (cx, cy, cz) = World::get_chunk_coord(x, y, z);
         let (dx, dy, dz) = (
             (x - cx * CHUNK_SIZE as i64) as u32,
@@ -76,7 +76,7 @@ impl World {
 
     /// Set data at position x,y,z
     /// Enventually create a new chunk if necessary
-    pub fn _set_data(&mut self, x: i64, y: i64, z: i64, data: u32) {
+    pub fn _set_data(&mut self, x: i64, y: i64, z: i64, data: u16) {
         let (cx, cy, cz) = World::get_chunk_coord(x, y, z);
         let (dx, dy, dz) = (
             (x - cx * CHUNK_SIZE as i64) as u32,
